+++ changelog.md
@@ -24,6 +24,7 @@
 - inactive stages don't error out on creation
 - fixed suspected bug where users logging in for the first time would be denied a restricted page even if they should have permission
 - fixed forge faction duplication bug where button did not trigger
+- new user profiles failed to save avatar correctly
 
 ## [1.13.3] - 2026-8-21 Post Spain Updates